+++ CHANGELOG.md
@@ -45,3 +45,16 @@
 - Changed the query editor layout
 - Support Grafana version 11
 - Drop support for Grafana 8.x and 9.x
+
+## 5.1.0
+
+- Add units and description to new format - issue #154
+- Fixed digital state - issue #159
+- Fixed summary data - issue #160
+- Fixed an error in recorded max number of points - issue #162
+
+- Updated the query editor layout
+- Added boundary type support in recorded values
+- Recognize partial usage of variables in elements
+- Added configuration to hide API errors in panel
+- Truncate time from grafana date time picker to seconds